test(frontend): consolidate route registry test coverage (#4413)
App.tsx and main.tsx already derive route mode through the shared
routes/registry.ts (via pageManifest.ts) rather than duplicating the
mapping — no source change needed there. Fold the useful assertions
from the stale pageManifest.test.tsx (which asserted a non-existent
`entry.menu` shape, so its checks were silently no-ops) into
pageManifest.test.ts, and add an explicit per-segment check that
deriveModeFromPathname (used at runtime by App.tsx) and
deriveBootstrapMode (used at bootstrap by main.tsx) agree for every
known route.
